refactor sensum code more generic
